src/backend/requirements.txt
# Core API and Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.7

# Async and HTTP
httpx>=0.25.2
aiofiles>=23.2.1
asyncpg>=0.29.0

# Data Processing and AI
anthropic>=0.40.0
langchain>=0.1.0
langchain-anthropic>=0.1.0

# Web Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.2
requests>=2.31.0

# Task Queue
celery>=5.3.4
redis>=5.0.1

# Monitoring and Logging
flower>=2.0.1

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0

# RAG Pipeline Dependencies
chromadb>=0.4.24
sentence-transformers>=2.6.1

# PDF Processing for MCP Orchestrator
pdfplumber>=0.10.0
PyPDF2>=3.0.1
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
